Sampling from Exponential Distribution

The srexp() function generates random samples from a Exponential Distribution using the STORS algorithm. It employs an optimized proposal distribution around the mode and Inverse Transform (IT) method for the tails.

Usage

srexp(n = 1, rate = 1, x = NULL)

srexp_custom(n = 1, x = NULL)

Arguments

n

Integer, length 1. Number of samples to draw.

rate

Numeric. is the rate parameter of the Exponential Distribution.

x

(optional) Numeric vector of length \(n\). If provided, this vector is over written in place to avoid any memory allocation.

Value

A numeric vector of length n containing samples from the Exponential Distribution with the specified rate.

NOTE: When the x parameter is specified, it is updated in-place with the simulation for performance reasons.

Details

The Exponential distribution has the probability density function (PDF): \(f(x | \lambda) = \lambda \exp(-\lambda x), \quad x \geq 0,\) where:

\(\lambda\)

is the rate parameter (\(\lambda > 0\)), which determines the rate of decay of the distribution.

The Exponential distribution is commonly used to model the time between independent events that occur at a constant average rate.

These two functions are for sampling using the STORS algorithm based on the proposal that has been constructed using srexp_optimize.

By default, srexp() samples from a standard Exponential Distribution rate = 1. The proposal distribution is pre-optimized at package load time using srexp_optimize() with steps = 4091, creating a scalable proposal centred around the mode.

If srexp() is called with custom rate parameter, the samples are generated from the standard Exponential Distribution, then scaled accordingly.

See also

Examples

# Generate 10 samples from the standard Exponential Distribution
samples <- srexp(10)
print(samples)
#>  [1] 2.0217216 0.6367446 0.2157311 0.2058001 0.9501565 2.0056431 0.6597147
#>  [8] 0.3001304 0.1086701 0.1408734

# Generate 10 samples using a pre-allocated vector
x <- numeric(10)
srexp(10, x = x)
#>  [1] 0.66569863 0.05568734 0.98998579 0.48884617 0.28460393 0.20524387
#>  [7] 0.48706766 0.73185619 0.34453423 1.45612172
print(x)
#>  [1] 0.66569863 0.05568734 0.98998579 0.48884617 0.28460393 0.20524387
#>  [7] 0.48706766 0.73185619 0.34453423 1.45612172

# Generate 10 samples from a Exponential Distribution with rate = 4
samples <- srexp(10, rate = 4)
print(samples)
#>  [1] 0.479206767 0.057603932 0.588815025 0.666546820 0.380122215 0.449768534
#>  [7] 0.009975796 0.647500648 0.021023598 0.168533075