Sampling from Normal Distribution

The srnorm() function generates random samples from a Normal distribution using the STORS algorithm. It employs an optimized proposal distribution around the mode and Adaptive Rejection Sampling (ARS) for the tails.

Usage

srnorm(n = 1, mean = 0, sd = 1, x = NULL)

srnorm_custom(n = 1, x = NULL)

Arguments

n

Integer, length 1. Number of samples to draw.

mean

Numeric. Mean parameter of the Normal distribution.

sd

Numeric. Standard deviation of the target Normal distribution.

x

(optional) Numeric vector of length \(n\). If provided, this vector is over written in place to avoid any memory allocation.

Value

A numeric vector of length n containing samples from the Normal distribution with the specified mean and sd.

NOTE: When the x parameter is specified, it is updated in-place with the simulation for performance reasons.

Details

The Normal distribution has the probability density function (PDF): \(f(x | \mu, \sigma) = \frac{1}{\sigma\sqrt{2\pi}} \exp\left(-\frac{(x - \mu)^2}{2\sigma^2}\right),\) where:

\(\mu\)

is the mean of the distribution, which determines the center of the bell curve.

\(\sigma\)

is the standard deviation, which controls the spread of the distribution (\(\sigma > 0\)).

These two functions are for sampling using the STORS algorithm based on the proposal that has been constructed using srnorm_optimize.

By default, srnorm() samples from a standard Normal distribution (mean = 0, sd = 1). The proposal distribution is pre-optimized at package load time using srnorm_optimize() with steps = 4091, creating a scalable proposal centered around the mode.

If srnorm() is called with custom mean or sd parameters, the samples are generated from the standard Normal distribution, then scaled and location shifted accordingly.

See also

Examples

# Generate 10 samples from the standard Normal distribution
samples <- srnorm(10)
print(samples)
#>  [1]  1.82162188 -1.54503316  0.66665189 -0.18628797 -1.13132568 -0.04051822
#>  [7] -0.31578765 -0.54573070  1.19908447 -1.07421382

# Generate 10 samples using a pre-allocated vector
x <- numeric(10)
srnorm(10, x = x)
#>  [1] -0.8836765  1.2722961 -0.8249442  1.4520775  0.3352823  1.1540284
#>  [7]  2.4333979 -0.6606458  0.3987386 -1.0249183
print(x)
#>  [1] -0.8836765  1.2722961 -0.8249442  1.4520775  0.3352823  1.1540284
#>  [7]  2.4333979 -0.6606458  0.3987386 -1.0249183

# Generate 10 samples from a Normal distribution with mean = 2 and sd = 3
samples <- srnorm(10, mean = 2, sd = 3)
print(samples)
#>  [1]  2.0166136  4.7186473 -0.8259597  3.0032599  2.6041036  1.3932055
#>  [7]  4.7839186  3.1892148  0.3062024  3.6714986