Sampling from Exponential Distribution

The srexp() function generates random samples from a Exponential Distribution using the STORS algorithm. It employs an optimized proposal distribution around the mode and Inverse Transform (IT) method for the tails.

Usage

srexp(n = 1, rate = 1, x = NULL)

srexp_custom(n = 1, x = NULL)

Arguments

n

Integer, length 1. Number of samples to draw.

rate

Numeric. is the rate parameter of the Exponential Distribution.

x

(optional) Numeric vector of length \(n\). If provided, this vector is over written in place to avoid any memory allocation.

Value

A numeric vector of length n containing samples from the Exponential Distribution with the specified rate.

NOTE: When the x parameter is specified, it is updated in-place with the simulation for performance reasons.

Details

The Exponential distribution has the probability density function (PDF): \(f(x | \lambda) = \lambda \exp(-\lambda x), \quad x \geq 0,\) where:

\(\lambda\)

is the rate parameter (\(\lambda > 0\)), which determines the rate of decay of the distribution.

The Exponential distribution is commonly used to model the time between independent events that occur at a constant average rate.

These two functions are for sampling using the STORS algorithm based on the proposal that has been constructed using srexp_optimize.

By default, srexp() samples from a standard Exponential Distribution rate = 1. The proposal distribution is pre-optimized at package load time using srexp_optimize() with steps = 4091, creating a scalable proposal centered around the mode.

If srexp() is called with custom rate parameter, the samples are generated from the standard Exponential Distribution, then scaled accordingly.

See also

Examples

# Generate 10 samples from the standard Exponential Distribution
samples <- srexp(10)
print(samples)
#>  [1] 0.5603843 0.2844683 0.4954779 2.1691440 0.1536847 0.1738814 0.6721279
#>  [8] 0.5504264 0.1415577 0.3971108

# Generate 10 samples using a pre-allocated vector
x <- numeric(10)
srexp(10, x = x)
#>  [1] 0.59366667 1.24779486 0.97384716 0.24309197 1.03282179 1.36348987
#>  [7] 0.32088562 2.05045399 0.17722475 0.02695628
print(x)
#>  [1] 0.59366667 1.24779486 0.97384716 0.24309197 1.03282179 1.36348987
#>  [7] 0.32088562 2.05045399 0.17722475 0.02695628

# Generate 10 samples from a Exponential Distribution with rate = 4
samples <- srexp(10, rate = 4)
print(samples)
#>  [1] 0.041349464 0.224742396 0.106423296 0.478133937 0.112295795 0.091989576
#>  [7] 1.295193902 0.006852274 0.740551723 0.124286375